backend/requirements.txt
fastapi
httpx
cachetools
google-generativeai
yfinance
python-dotenv
pandas
